web-app/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
mysql-connector-python==8.0.25
prometheus_client==0.14.1
requests==2.26.0
